feat: add deims id parsing for dataregistry
While implementing this hardcoded assigning of the deims id, several
approaches were considered - we can try to download the file, and
finding the site's deims id. We can even leverage of an endpoint:
https://deims.org/search_api_autocomplete, but that is likely meant only
for deims's UI for searching. Thus, this approach was chosen for now,
and better solution might be implemented in the future if needed.