feat: add sync of one record
This commit adds support for harvesting of one file. This will help in
supporting of refresh of one record after a rule is added, as well as it
might be possible to refresh one specific file. This also tries to
detect a mismatch between savd source_url and the one used.